DB-12: Updated interactive.go and Added  utils.go
